Enable SegmentCompleteionIntegrationTest
This test was not running because of a typo in the class name.
Also, re-organized the test so that it does not have to override
methods from upper level tests each time new tests are added.
