+++ pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkRealtimeConsumptionSpeed.java
@@ -82,7 +82,7 @@ public class BenchmarkRealtimeConsumptionSpeed extends RealtimeClusterIntegratio
     startServer();
 
     // Create realtime table
-    setUpTable(avroFiles.get(0));
+    setUpRealtimeTable(avroFiles.get(0));
 
     // Wait a couple of seconds for all Helix state transitions to happen
     Uninterruptibles.sleepUninterruptibly(5, TimeUnit.SECONDS);
